requirements.txt
# === Core ===
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18

# === Data ===
pandas>=2.2.0
numpy>=1.26.0

# === ML ===
scikit-learn>=1.6.0
prophet>=1.1.6

# === NLP / LLM ===
openai>=1.60.0

# === Visualization ===
plotly>=5.24.0
streamlit>=1.41.0

# === Reports ===
fpdf2>=2.8.0

# === Testing ===
pytest>=8.3.0
httpx>=0.28.0
pytest-cov>=6.0.0
pytest-asyncio>=0.25.0

# === Utilities ===
python-dotenv>=1.0.1
chardet>=5.2.0
